// src/enums/cst-pis-cofins-saida.enum.ts
export enum CstPisCofinsSaidaEnum {
  CST_PIS_COFINS_SAIDA_01 = "01",
  CST_PIS_COFINS_SAIDA_02 = "02",
  CST_PIS_COFINS_SAIDA_03 = "03",
  CST_PIS_COFINS_SAIDA_04 = "04",
  CST_PIS_COFINS_SAIDA_05 = "05",
  CST_PIS_COFINS_SAIDA_06 = "06",
  CST_PIS_COFINS_SAIDA_07 = "07",
  CST_PIS_COFINS_SAIDA_08 = "08",
  CST_PIS_COFINS_SAIDA_09 = "09",
  CST_PIS_COFINS_SAIDA_49 = "49",
}